# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EBookLibAPI",
    "description": "API thư viện điện tử",
    "version": "v1"
  },
  "paths": {
...
    "/api/Borrow/register": {
      "post": {
        "tags": [
          "Borrow"
        ],
        "summary": "Đăng ký mượn sách",
        "requestBody": {
          "description": "IdBanDoc: id Bạn đọc, IdXepGia: id Xếp giá, Ghi chú",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRegistrationRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRegistrationRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowRegistrationRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BorrowRegistrationDtoApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc đã hết hạn"
          },
          "403": {
            "description": "Forbidden - Không có quyền truy cập tài nguyên này"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Borrow/approve": {
      "post": {
        "tags": [
          "Borrow"
        ],
        "summary": "Duyệt yêu cầu mượn (chỉ tài khoản Admin/Thủ thư)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowApprovalRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowApprovalRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowApprovalRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BorrowDetailDtoApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc đã hết hạn"
          },
          "403": {
            "description": "Forbidden - Không có quyền truy cập tài nguyên này"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/News/{id}": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Lấy tin tức chi tiết",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id Tin tức",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetailDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "News"
        ],
        "summary": "Cập nhật bình luận",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id Bình luận",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Nội dung cmt",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc đã hết hạn"
          },
          "403": {
            "description": "Forbidden - Không có quyền truy cập tài nguyên này"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "Xóa bình luận",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id Bình luận",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc đã hết hạn"
          },
          "403": {
            "description": "Forbidden - Không có quyền truy cập tài nguyên này"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/News": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Lấy danh sách cmt của bài tin",
        "parameters": [
          {
            "name": "tinTucId",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Thêm mới comment",
        "requestBody": {
          "description": "TinTucId: id Tin tức, ParentId: Id bình luận cha (null nết là root), ReplyToId: trả lời cmt - id của comment trả lời",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized - Token không hợp lệ hoặc đã hết hạn"
          },
          "403": {
            "description": "Forbidden - Không có quyền truy cập tài nguyên này"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BorrowApprovalRequestDto": {
        "type": "object",
        "properties": {
          "idYeuCauMuon": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BorrowDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "idMuonTra": {
            "type": "string",
            "nullable": true
          },
          "idBanDoc": {
            "type": "string",
            "nullable": true
          },
          "tenBanDoc": {
            "type": "string",
            "nullable": true
          },
          "soThe": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "soDienThoai": {
            "type": "string",
            "nullable": true
          },
          "ngayHetHan": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idAnPham": {
            "type": "string",
            "nullable": true
          },
          "tenAnPham": {
            "type": "string",
            "nullable": true
          },
          "tacGia": {
            "type": "string",
            "nullable": true
          },
          "maDKCB": {
            "type": "string",
            "nullable": true
          },
          "ngayDangKy": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ngayMuon": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ngayTra": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ngayTraDuKien": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "phiMuon": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "trangThai": {
            "type": "string",
            "nullable": true
          },
          "tenTrangThai": {
            "type": "string",
            "nullable": true
          },
          "ngoaiHanNgach": {
            "type": "boolean",
            "nullable": true
          },
          "idXepGiaDetail": {
            "type": "string",
            "nullable": true
          },
          "ghiChu": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BorrowDetailDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "errorId": {
            "type": "string",
            "nullable": true
          },
          "appId": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/BorrowDetailDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BorrowRegistrationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "idBanDoc": {
            "type": "string",
            "nullable": true
          },
          "tenBanDoc": {
            "type": "string",
            "nullable": true
          },
          "soThe": {
            "type": "string",
            "nullable": true
          },
          "idXepGia": {
            "type": "string",
            "nullable": true
          },
          "maDKCB": {
            "type": "string",
            "nullable": true
          },
          "tenAnPham": {
            "type": "string",
            "nullable": true
          },
          "tacGia": {
            "type": "string",
            "nullable": true
          },
          "ngayDangKy": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ngayTraDuKien": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "phiMuon": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "trangThai": {
            "type": "string",
            "nullable": true
          },
          "tenTrangThai": {
            "type": "string",
            "nullable": true
          },
          "maDangKyCaBiet": {
            "type": "string",
            "nullable": true
          },
          "ngayTao": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BorrowRegistrationDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "errorId": {
            "type": "string",
            "nullable": true
          },
          "appId": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/BorrowRegistrationDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BorrowRegistrationRequestDto": {
        "type": "object",
        "properties": {
          "idBanDoc": {
            "type": "string",
            "nullable": true
          },
          "idXepGia": {
            "type": "string",
            "nullable": true
          },
          "ngayDangKy": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "maDangKyCaBiet": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CommentRequest": {
        "required": [
          "noiDung",
          "tinTucId"
        ],
        "type": "object",
        "properties": {
          "tinTucId": {
            "minLength": 1,
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "replyToId": {
            "type": "string",
            "nullable": true
          },
          "noiDung": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CommentUpdateRequest": {
        "required": [
          "noiDung"
        ],
        "type": "object",
        "properties": {
          "noiDung": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "FileNewsInfo": {
        "type": "object",
        "properties": {
          "TEN": {
            "type": "string",
            "nullable": true
          },
          "DUONG_DAN": {
            "type": "string",
            "nullable": true
          },
          "DUNG_LUONG": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "NewsDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "tieuDe": {
            "type": "string",
            "nullable": true
          },
          "anhDaiDien": {
            "type": "string",
            "nullable": true
          },
          "moTa": {
            "type": "string",
            "nullable": true
          },
          "luotXem": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ngayDang": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "tenNhomTinTuc": {
            "type": "string",
            "nullable": true
          },
          "noiDung": {
            "type": "string",
            "nullable": true
          },
          "tepDinhKem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileNewsInfo"
            },
            "nullable": true
          },
          "danhSachTinLienQuan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewsDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewsDetailDtoApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "errorId": {
            "type": "string",
            "nullable": true
          },
          "appId": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/NewsDetailDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "tieuDe": {
            "type": "string",
            "nullable": true
          },
          "anhDaiDien": {
            "type": "string",
            "nullable": true
          },
          "moTa": {
            "type": "string",
            "nullable": true
          },
          "luotXem": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ngayDang": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Nhập JWT token vào đây. Ví dụ: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}